Add development seeds, rake task, and logging
The solution I eventually decided on for performance logging/tracking I picked
as the most maintainable and that provides some flexibility in the development
environment.

Instead of creating a staging server, a developer looking to test code can
seed the development db with `rake db:seed` (as usual), a process that will
take about 30 min to load 12000 companies and 30000 jobs. While this is lower
than the acceptance criteria, it still is an increase on what we have in
production currently and can be bumped up as our app gets bigger. Because of the
time to seed the db, I wasn't sure that any more records would be feasible for
developers shifting onto this project.

However, this allows a developer to use the new rake task `rake load_script:run`
to hit any endpoint in rapid succession, generating data to `logs/request.log`.
Additionally, we have access to New Relic data at the endpoint `localhost:3000/newrelic`.

Currently the rake task is only hitting the index and a random job listing, but
anyone can add new endpoints easily in `lib/load_script/job_endpoints` using
plain ol' Capybara.

While we may still need to decide on a better solution for viewing production
performance data, this PR gives some flexibility