Add proper capitalization for tech
Created a method in tech serializer which checks if a given tech is included in an array of unusually capitalized technology and returns what it should be, or if not included, upcases the technology.

Currently this just affects JavaScript, but this could be relevant should we ever include something like "RSpec" or acronyms. For those cases, they will need to be added to the hash in the serializer as well.

I've also added a quick test for this.